+++ dash/card-upload-act.php
@@ -26,25 +26,25 @@
         }
 
         // Apagando todas as sessões desse grupo
-        $sql = sprintf('delete from car_study_session where stud_id in (select stud_id from car_study where deck_id = %d)', $deck_id, $user_id);
+        $sql = sprintf('delete from car_study_session where stud_id in (select stud_id from car_study where deck_id = %d and user_id = %d)', $deck_id, $user_id);
 
         $result = $mysqli->query($sql);
 
-        if (!$result) error_log($mysqli->sqlstate . ' - ' . $mysqli->error); throw new Exception('error.db');
+        if (!$result) { error_log($mysqli->sqlstate . ' - ' . $mysqli->error); throw new Exception('error.db'); }
 
         // Apagando todos os estudos desse grupo
-        $sql = sprintf('delete from car_study where deck_id = %d', $deck_id, $user_id);
+        $sql = sprintf('delete from car_study where deck_id = %d and user_id = %d', $deck_id, $user_id);
 
         $result = $mysqli->query($sql);
 
-        if (!$result) error_log($mysqli->sqlstate . ' - ' . $mysqli->error); throw new Exception('error.db');
+        if (!$result) { error_log($mysqli->sqlstate . ' - ' . $mysqli->error); throw new Exception('error.db'); }
 
         // Apagando todos os cartões desse estudo
         $sql = sprintf('delete from car_card where deck_id = %d and user_id = %d', $deck_id, $user_id);
 
         $result = $mysqli->query($sql);
 
-        if (!$result) error_log($mysqli->sqlstate . ' - ' . $mysqli->error); throw new Exception('error.db');
+        if (!$result) { error_log($mysqli->sqlstate . ' - ' . $mysqli->error); throw new Exception('error.db'); }
 
         if (isset($_FILES['input_file']) && $_FILES['input_file']['error'] === UPLOAD_ERR_OK) {
             $file = $_FILES['input_file']['tmp_name'];
@@ -92,7 +92,7 @@
 
                             $result = $mysqli->query($sql);
 
-                            if (!$result) error_log($mysqli->sqlstate . ' - ' . $mysqli->error); throw new Exception('error.db');
+                            if (!$result) { error_log($mysqli->sqlstate . ' - ' . $mysqli->error); throw new Exception('error.db'); }
                         } else {
                             $line_import = true;
                         }
